Modify the libabfmulti API to make Sections an unsigned short. This corresponds to the API of the normal libabf. Also add title, author and time to the libabfmulti header file. We still need to do our WriteHeader function.
git-svn-id: svn://mosedal.net/abf/code/trunk@610 3c19b6e9-aa5f-dd11-88a4-001302d38fef

+++ libabfmulti/libabfmulti.cpp
@@ -10,7 +10,7 @@ namespace ABFMulti {
 using std::ostringstream;
 using std::cout;
 using std::endl;
-void AbfEncoder::SetNumParts(unsigned int NumSections) {
+void AbfEncoder::SetNumParts(unsigned short NumSections) {
 if (AbfParts != nullptr) {
 	delete[] AbfParts; // Free the memory
 	AbfParts = nullptr;
